+++ tests/shave.test.ts
@@ -40,6 +40,7 @@ function baseState(o: Partial<PromptFuzzState> = {}): PromptFuzzState {
     dailyLog: {},
     statusViewCount: 0,
     quietHours: null,
+    currentSession: null,
     ...o,
   };
 }